% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot_functions.R
\name{categorize_and_plot_genes}
\alias{categorize_and_plot_genes}
\title{Categorize and Plot Genes}
\usage{
categorize_and_plot_genes(
  string_results,
  pubmed_search_results,
  file_directory,
  threshold_percentage = 20
)
}
\arguments{
\item{string_results}{Data frame with STRING metrics.}

\item{pubmed_search_results}{Data frame with PubMed search results.}

\item{file_directory}{Directory for saving the output plot.}

\item{threshold_percentage}{Percentage threshold for ranking (default is 20\%).}
}
\description{
Categorize genes and create a scatter plot of Connectivity Rank vs. PubMed Rank.
}
