% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descide_master_functions.R
\name{descide}
\alias{descide}
\title{Run the Entire Analysis}
\usage{
descide(
  genes_list,
  terms_list,
  file_directory,
  export_format = "csv",
  threshold_percentage = 20,
  species = 9606,
  network_type = "full",
  score_threshold = 400,
  rank_method = "weighted",
  export = FALSE
)
}
\arguments{
\item{genes_list}{A list of genes.}

\item{terms_list}{A list of terms.}

\item{file_directory}{Directory for saving the output files.}

\item{export_format}{Format for export, either "csv", "tsv", or "excel".}

\item{threshold_percentage}{Percentage threshold for ranking (default is 20\%).}

\item{species}{The NCBI taxon ID of the species. Defaults to 9606 (Homo sapiens).}

\item{network_type}{The type of network to use, either "full" or "physical". Defaults to "full".}

\item{score_threshold}{The minimum score threshold for interactions. Defaults to 400.}

\item{rank_method}{The method to rank results, either "weighted" or "total". Defaults to "weighted".}

\item{export}{Logical indicating whether to export the results. Defaults to FALSE.}
}
\value{
A list containing the PubMed search results, STRING results, and combined summary path.
}
\description{
Run the entire analysis pipeline including PubMed search, STRING database search, and plotting.
}
