% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot_functions.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Plot STRING Interactions}
\usage{
plot_clustering(
  string_results,
  file_directory,
  current_date = format(Sys.Date(), "\%m.\%d.\%Y")
)
}
\arguments{
\item{string_results}{Data frame with STRING metrics.}

\item{file_directory}{Directory for saving the output plot.}

\item{current_date}{Current date for file naming.}
}
\description{
Plot STRING interactions degree vs. clustering.
}
