% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(pubmed_search_results, file_directory)
}
\arguments{
\item{pubmed_search_results}{A data frame containing raw search results with genes and terms.}

\item{file_directory}{Directory for saving the output plot.}
}
\description{
Create and save a heatmap of the PubMed search results.
}
