% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb_functions.R
\name{plot_string_network}
\alias{plot_string_network}
\title{Plot STRING Network}
\usage{
plot_string_network(string_db, string_ids, file_directory, current_date)
}
\arguments{
\item{string_db}{A STRINGdb object.}

\item{string_ids}{A list of STRING IDs.}

\item{file_directory}{Directory for saving the output plot.}

\item{current_date}{Current date for file naming.}
}
\description{
Plot STRING network interactions using STRINGdb.
}
