% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb_functions.R
\name{search_string_db}
\alias{search_string_db}
\title{Search STRING Database}
\usage{
search_string_db(genes_list)
}
\arguments{
\item{genes_list}{A list of genes.}
}
\value{
A list containing metrics, STRINGdb object, and STRING IDs.
}
\description{
Search the STRING database for protein interactions.
}
