% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_and_plot_functions.R
\name{combine_summary}
\alias{combine_summary}
\title{Combine PubMed and STRING Metrics}
\usage{
combine_summary(
  pubmed_search_results,
  string_results,
  file_directory,
  export_format = "csv",
  export = FALSE
)
}
\arguments{
\item{pubmed_search_results}{Data frame with PubMed search results.}

\item{string_results}{Data frame with STRING metrics.}

\item{file_directory}{Directory for saving the output summary.}

\item{export_format}{Format for export, either "csv", "tsv", or "excel".}

\item{export}{Logical indicating whether to export the summary. Defaults to FALSE.}
}
\value{
A data frame with combined summary.
}
\description{
Combine PubMed search summary and STRING gene metrics.
}
