% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descide_master_functions.R
\name{descide}
\alias{descide}
\title{Run the Entire Analysis}
\usage{
descide(
  genes_list,
  terms_list,
  file_directory,
  export_format = "csv",
  threshold_percentage = 20
)
}
\arguments{
\item{genes_list}{A list of genes.}

\item{terms_list}{A list of terms.}

\item{file_directory}{Directory for saving the output files.}

\item{export_format}{Format for export, either "csv", "tsv", or "excel".}

\item{threshold_percentage}{Percentage threshold for ranking (default is 20\%).}
}
\description{
Run the entire analysis pipeline including PubMed search, STRING database search, and plotting.
}
