% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb_functions.R
\name{plot_clustering}
\alias{plot_clustering}
\title{Plot STRING Interactions}
\usage{
plot_clustering(string_results, file_directory)
}
\arguments{
\item{string_results}{Data frame with STRING metrics.}

\item{file_directory}{Directory for saving the output plot.}
}
\description{
Plot STRING interactions degree vs. clustering.
}
