% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(pubmed_search_results, file_directory, export = FALSE)
}
\arguments{
\item{pubmed_search_results}{A data frame containing raw search results with genes and terms.}

\item{file_directory}{Directory for saving the output plot.}

\item{export}{Logical indicating whether to export the plot. Defaults to FALSE.}
}
\description{
Create and save (optional) a heatmap of the PubMed search results.
}
