% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_functions.R
\name{rank_search_results}
\alias{rank_search_results}
\title{Rank Search Results}
\usage{
rank_search_results(data, terms_list, rank_method = "weighted")
}
\arguments{
\item{data}{A data frame containing search results.}

\item{terms_list}{A list of terms used in the search.}

\item{rank_method}{The method to rank results, either "weighted" or "total".}
}
\value{
A data frame with ranked search results.
}
\description{
Rank search results based on a chosen method.
}
