% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_functions.R
\name{single_pubmed_search}
\alias{single_pubmed_search}
\title{Search PubMed}
\usage{
single_pubmed_search(gene, term)
}
\arguments{
\item{gene}{A character string representing the gene symbol.}

\item{term}{A character string representing the search term.}
}
\value{
An integer representing the number of PubMed articles found from the search query in PubMed.
}
\description{
Perform a PubMed search for a given gene and term.
}
\examples{
# Perform a PubMed search for gene 'TP53' with term 'cancer'
result <- single_pubmed_search("TP53", "cancer")
print(result)
}
